requirements.txt
peewee~=3.18.1
langchain-core~=0.3.62
fastapi~=0.115.12
qdrant-client~=1.14.2
fastembed
python-dotenv~=1.0.1
uvicorn~=0.34.2
langchain-huggingface
langchain-openai~=0.3.8
langchain-qdrant
langchain-experimental~=0.3.4
pydantic~=2.10.6
langgraph~=0.3.10
pdf2image~=1.17.0
opencv-python~=4.11.0.86
torch~=2.7.1
pillow~=11.2.1
pathlib~=1.0.1
langchain_google_genai
gunicorn
sentence-transformers
PyPDF2
python-multipart
